// src/modules/activty-logs/enums/activity-log-event-type.ts
export enum ActivityLogEventType {
  TASK_STARTED = 'TASK_STARTED',
  TASK_LAST_ACCESSED = 'TASK_LAST_ACCESSED',
  TASK_SUBMITTED = 'TASK_SUBMITTED',
  TASK_COMPLETED = 'TASK_COMPLETED',
  TASK_GRADED = 'TASK_GRADED',
  SHARE_TASK = 'SHARE_TASK',
  JOIN_CLASS = 'JOIN_CLASS',
  LEAVE_CLASS = 'LEAVE_CLASS',
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
}